+++ testFrameworkServer/src/ch/hsr/objectCaching/testFrameworkServer/TestCaseFactory.java
@@ -57,7 +57,7 @@ public class TestCaseFactory
 		    	  if(stax.hasName() && stax.getName().toString().equals("Account") && stax.isStartElement())
 		    	  {
 		    		  newAccount = new AccountImpl();
-		    		  newAccount.setBalance(Integer.valueOf(stax.getAttributeValue(0)));
+		    		  newAccount.setBalance(Double.valueOf(stax.getAttributeValue(0)));
 		    		  accounts.add(newAccount);
 		    	  }
